+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -214,5 +214,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-18T06:46:59.426Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-18T11:10:42.158Z")
     r"""The date and time the resource was last updated"""