+++ packages/orq-rc/src/orq_ai_sdk/models/listagentsop.py
@@ -224,7 +224,7 @@ class ListAgentsKnowledgeBases(BaseModel):
     configuration: ListAgentsConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K5DS7B50TSZVG673GHJ1DK5J"
+    id: Optional[str] = "01K5E8A6RZT1PCVK58XFAA34WS"
     r"""The id of the resource"""