+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -127,7 +127,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5DS7B98M0KSDMY5CQMXD4SK"
+        "01K5E8A6X18R7ZVH7X20VHEYGS"
     )
     r"""The id of the resource"""