+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5DS7B99FQP1NTC6C59HDJBC"
+        "01K5E8A6X2VQQPK7A8V82W82FN"
     )
     r"""The id of the resource"""