+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -433,7 +433,7 @@ class Eleven(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5DS7B5TD01V1SFT51HDT1XG"
+        "01K5E8A6SKDM6WDDRGVD1YNSN1"
     )
 
     status: Optional[ToolsStatus] = "live"