+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -458,7 +458,7 @@ class Tools11(BaseModel):
     http: ToolsHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5DS7B6171EJH0CX4QFMQMYX"
+        "01K5E8A6T4BF7JCDEVKBT31XV7"
     )
 
     status: Optional[StreamRunAgentToolsStatus] = "live"