+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1263,5 +1263,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-18T06:46:59.426Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-18T11:10:42.158Z")
     r"""The date and time the resource was last updated"""