+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5DS7B9BQG0ACYNW77QBHN4M"
+        "01K5E8A6X4TEMWM0SWWJ7BG1M2"
     )
     r"""The id of the resource"""